Move the struct resolve test to wgslGenerator.test.ts
